requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
transformers==4.36.2
torch==2.1.2
openai==1.6.1
langchain==0.0.350
langchain-openai==0.0.2
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Web Dashboard Dependencies
streamlit==1.32.0
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4.post3
alembic==1.13.1
